Add files via upload
The UOL coursework is divided into 2 parts. 
The 1st part is to the Markov Chain Monte Carlo algorithm and the Metropolis-Hastings algorithm. The aim is to simulate random numbers for the distribution with probability density function.
The second part is to derive and present the airport data sets from 1997 to 2000 of the different air flights in America. The 2009 ASA Statistical Computing and Graphics Data Expo consisted of flight arrival and departure details for all commercial flights on major carriers within the USA from October 1987 to April 2008. This is a large dataset; there are nearly 120 million records in total, and it takes up 1.6 gigabytes of space when compressed and 12 gigabytes when uncompressed. The complete dataset, along with supplementary information and variable descriptions, can be downloaded from the Harvard Dataverse at https://doi.org/10.7910/DVN/HG7NV7
When using the R Markdown and Python files for Part 2, it is recommended to create your own R markdown and Python files. Then create your own working directories for the Part 2 R and Python files you created and use the GitHub files here to add into your files. This way the newly created coding files will not have any working directory issues.